try to specify more types explicitly
martin@labolg ~/stage/cl-cpp-generator2/example/33_copernicus_gtk/source $ make
c++ -c -o vis_04_decode_packet.o vis_04_decode_packet.cpp -Os -g3 -ggdb -Wall -Wextra -Werror -Wshadow -Wdouble-promotion -fno-common -Wconversion -ffunction-sections -fdata-sections -Wno-error=unused-variable -pipe `pkg-config gtkmm-3.0 --cflags` -march=native -mtune=native
vis_04_decode_packet.cpp: In function 'int init_decode_packet(int, std::complex<float>*)':
vis_04_decode_packet.cpp:555:38: error: catching polymorphic type 'class std::out_of_range' by value [-Werror=catch-value=]
  555 |           } catch (std::out_of_range e) {
      |                                      ^
vis_04_decode_packet.cpp:586:49: error: conversion from 'float' to 'std::array<const float, 4>::size_type' {aka 'long unsigned int'} may change value [-Werror=float-conversion]
  586 |             v = ((symbol_sign) * (table_nrl0.at(mcode)) * (table_sf.at(thidx)));
      |                                                 ^~~~~
vis_04_decode_packet.cpp:587:38: error: catching polymorphic type 'class std::out_of_range' by value [-Werror=catch-value=

+++ example/33_copernicus_gtk/source/utils.h
@@ -25,15 +25,15 @@ inline bool get_sequential_bit(sequential_bit_t *seq_state) {
   }
   return res;
 }
-inline int get_threshold_index(sequential_bit_t *s) {
-  return ((((0x80) * (get_sequential_bit(s)))) +
-          (((0x40) * (get_sequential_bit(s)))) +
-          (((0x20) * (get_sequential_bit(s)))) +
-          (((0x10) * (get_sequential_bit(s)))) +
-          (((0x8) * (get_sequential_bit(s)))) +
-          (((0x4) * (get_sequential_bit(s)))) +
-          (((0x2) * (get_sequential_bit(s)))) +
-          (((0x1) * (get_sequential_bit(s)))));
+inline uint8_t get_threshold_index(sequential_bit_t *s) {
+  return static_cast<uint8_t>(((((0x80) * (get_sequential_bit(s)))) +
+                               (((0x40) * (get_sequential_bit(s)))) +
+                               (((0x20) * (get_sequential_bit(s)))) +
+                               (((0x10) * (get_sequential_bit(s)))) +
+                               (((0x8) * (get_sequential_bit(s)))) +
+                               (((0x4) * (get_sequential_bit(s)))) +
+                               (((0x2) * (get_sequential_bit(s)))) +
+                               (((0x1) * (get_sequential_bit(s))))));
 }
 
 #endif
